requirements/production.txt
# Core
# ------------------------------------------------------------------------------
fastapi==0.61.1
pydantic==1.6.1
uvicorn==0.11.8
orjson==3.3.1

# Database
# ------------------------------------------------------------------------------
tortoise-orm==0.16.14
aerich==0.2.2

# Sentry
# ------------------------------------------------------------------------------
sentry-sdk==0.17.3
